# swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: "trentose2jan12-esame-api"
  description: Api per l'esame
  termsOfService: http://swagger.io/terms/
  contact:
    name: Francesco Bruschetti
    email: francesco.bruschetti@studenti.unitn.it
    url: http://madskristensen.net
  license:
    name: MIT
    url: http://github.com/gruntjs/grunt/blob/master/LICENSE-MIT
host: trentose2jan12-bruschetti.herokuapp.com
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /astronauts:
    get:
      description: |
        Returns all astronauts and their details
      responses:
        200:
          description: astronauts deatails
          schema:
            type: array
            items:
              $ref: '#/definitions/Astronauts'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      description: Creates a new astronaut profile
      operationId: addAstronaut
      parameters:
        - name: astronaut
          in: body
          description: astronaut details
          required: true
          schema:
            $ref: '#/definitions/NewAstronaut'
      responses:
        200:
          description: astronaut created
          schema:
            $ref: '#/definitions/Astronaut'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /astronauts/{id}:
    get:
      description: Returns the astronaut with the specified ID
      operationId: getAstronautById
      parameters:
        - name: id
          in: path
          description: ID of astronaut to fetch
          required: true
          type: string
      responses:
        200:
          description: Searched astronaut data
          schema:
            $ref: '#/definitions/Astronaut'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
            
    put:
      description: Updates an existing astronaut
      operationId: updateAstronaut
      parameters:
        - name: id
          in: path
          description: ID of astronaut to edit or create
          required: true
          type: string
          
        - name: astronaut
          in: body
          description: astronaut
          required: true
          schema:
            $ref: '#/definitions/Astronauts'
      responses:
        202:
          description: astronaut update response
          schema:
            $ref: '#/definitions/Astronauts'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    
definitions: # defined the structure of how the server answer to this actions
  Astronaut:
    allOf:
      - $ref: '#/definitions/NewAstronaut'
      - required:
        - astronautID
        properties:
          astronautID:
            type: string
  NewAstronaut:
    required:
      - firstName
      - lastName
      - astronautID
      - astronoutResult
    properties:
      firstName:
        type: string
      lastName:
        type: string
      astronoutResult:
        type: object
      status:
        type: string
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string